fixed jQuery calls and "use strict" in functional scope
Removed the last occurences of $('...') instead of jQuery() because of framework compatibility.

Also moved the "resetField" function into the onready function. The "use strict" is also moved into the ready function because this should not set global. Some extension could use non strict variable declarations and so this would break because we would declare them as wrong in global. "use strict" is now in functional scope.